# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ServiceFlow SDM API
  version: 1.0.0
  description: |
    Service Delivery Manager backend API for MGM service delivery operations.

    ## Authentication
    All protected endpoints require a Bearer JWT access token obtained from
    `POST /api/v1/auth/login` or `POST /api/v1/auth/sso/exchange`.

    ## Roles
    Roles in order of privilege: `admin > sdm > manager > tam > csm > engineer > viewer`

    `viewer` has read-only access scoped to their assigned properties/incidents.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    IntegrationSource:
      type: object
      required: [source_name]
      properties:
        source_name:
          type: string
        enabled:
          type: boolean
        base_url:
          type: string
          nullable: true
        auth_type:
          type: string
          nullable: true
        schedule:
          type: string
          nullable: true
          example: "*/15 * * * *"
        credentials_ref:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        updated_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SyncResult:
      type: object
      required: [processed]
      properties:
        processed:
          type: integer
        status:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
    UserPropertyAssignment:
      type: object
      required: [property_id, name, property_type]
      properties:
        property_id:
          type: string
          format: uuid
        name:
          type: string
        property_type:
          type: string
          enum: [hotel, casino, resort, venue]
paths:
...
  /admin/sources:
    get:
      summary: List all integration source configs
      description: Requires role `admin` or `sdm`.
      tags: [Admin]
      responses:
        "200":
          description: Array of source configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntegrationSource"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/sources/{sourceName}:
    put:
      summary: Update an integration source config
      description: Requires role `admin`.
      tags: [Admin]
      parameters:
        - name: sourceName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                baseUrl:
                  type: string
                  format: uri
                authType:
                  type: string
                schedule:
                  type: string
                credentialsRef:
                  type: string
                notes:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          description: Updated source config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationSource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/sources/{sourceName}/test:
    post:
      summary: Test / manually run a source sync
      description: Requires role `admin` or `sdm`.
      tags: [Admin]
      parameters:
        - name: sourceName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── User Property Assignments ──────────────────────────────────────────────
  /admin/users/{userId}/properties:
    get:
      summary: List properties assigned to a user
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Array of property assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPropertyAssignment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: Assign properties to a user (additive)
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propertyIds]
              properties:
                propertyIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        "201":
          description: Properties assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                  assigned:
                    type: array
                    items:
                      type: string
                      format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    put:
      summary: Replace all property assignments for a user
      description: Requires role `admin` only. Deletes existing assignments and sets new ones.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [propertyIds]
              properties:
                propertyIds:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        "200":
          description: Properties replaced
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                  properties:
                    type: array
                    items:
                      type: string
                      format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{userId}/properties/{propertyId}:
    delete:
      summary: Remove one property assignment from a user
      description: Requires role `admin`, `sdm`, or `manager`.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Assignment removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Shared parameters & responses ──────────────────────────────────────────
...